requirements/base.txt
beautifulsoup4==4.7.1
bleach==2.1.4
certifi==2019.6.16
chardet==3.0.4
Django==2.1.9
django-app-namespace-template-loader==0.4.1
django-blog-zinnia==0.20
django-ckeditor==5.7.1
django-classy-tags==0.9.0
django-colorful==1.3
django-contact-form==1.7
django-contrib-comments==1.9.1
django-debug-toolbar==2.0
django-environ==0.4.5
django-filter==2.1.0
django-graphos==0.3.41
django-guardian==2.0.0
django-inline-actions==2.2.0
django-inline-svg==0.1.1
django-js-asset==1.2.2
django-mptt==0.9.1
django-nyt==1.1.3
django-redis==4.10.0
django-sekizai==1.0.0
django-tagging==0.4.6
django-userena-ce==4.1.1
django-webpack-loader==0.6.0
django-xmlrpc==0.1.8
djangorestframework==3.9.4
dry-rest-permissions==0.1.10
easy-thumbnails==2.6
html2text==2018.1.9
html5lib==1.0.1
idna==2.8
Markdown==2.6.11
mots-vides==2015.5.11
oauthlib==3.0.1
olefile==0.46
Pillow==7.1.1
psycopg2==2.8.6
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
regex==2019.6.8
requests==2.22.0
requests-oauthlib==1.2.0
rules==2.0.1
simplejson==3.16.0
six==1.12.0
sorl-thumbnail==12.5.0
soupsieve==1.9.2
sqlparse==0.3.0
urllib3==1.25.3
webencodings==0.5.1
wiki==0.4.4
zinnia-wysiwyg-ckeditor==1.3
